services/email-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
Jinja2==3.1.2
smtplib-ssl==1.0.0 